+++ CHANGELOG.md
@@ -3103,3 +3103,4 @@
 - [2026-07-23 22:57] test(perf): add unit tests for perf business logic
 - [2026-07-23 22:58] fix(auth): correct layout constraints and overflow in auth
 - [2026-07-24 09:46] feat(ui): implement responsive ui view component
+- [2026-07-24 10:18] test(network): implement widget integration tests for network